+++ src/index.ts
@@ -11,7 +11,7 @@ export * as BodyParserMiddleware from "./middlewares/body-parser.js";
 export * as CorsMiddleware from "./middlewares/CorsMiddleware.js";
 export * as CurrentPageNumberMiddleware from "./middlewares/CurrentPageNumberMiddleware.js";
 export * as ForceSslMiddleware from "./middlewares/ForceSslMiddleware.js";
-export * as LogRequestMiddleware from "./middlewares/log-request.js";
+export * as LogRequestMiddleware from "./middlewares/LogRequestMiddleware.js";
 export * as RouterMiddleware from "./middlewares/router.js";
 export * as StaticMiddleware from "./middlewares/static.js";