+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * refactor: made CorsMiddleware return its options
 * refactor: made CurrentPageNumberMiddleware return its options
 * refactor: made ForceSslMiddleware return its options
+* refactor: made LogRequestMiddleware return its options
 
 ## 2.1.0